// legacy-src/registry/types.ts
export interface TypeInfo {
  type: TypeRef;
  isSingleton: boolean;
}
// ...
export interface Analyzer {
  error(message: string): void;
  warning(message: string): void;
  resolveType(typeName: string): TypeRef;
}